fix GGC signed pointer warnings and switch volume_id to stdint
Solaris uses volume_id now and they fiddled around with configure scripts
to map the linux kernel int types. Adding the types locally to volume_id
breaks the klibc build, so just switch to these ugly types and forget it. :)

+++ extras/volume_id/volume_id/util.h
@@ -37,22 +37,22 @@
 #define SEEK_BUFFER_SIZE			0x10000
 
 /* probe volume for all known filesystems in specific order */
-#define bswap16(x) (__u16)((((__u16)(x) & 0x00ffu) << 8) | \
-			   (((__u16)(x) & 0xff00u) >> 8))
+#define bswap16(x) (uint16_t)	((((uint16_t)(x) & 0x00ffu) << 8) | \
+				(((uint16_t)(x) & 0xff00u) >> 8))
 
-#define bswap32(x) (__u32)((((__u32)(x) & 0xff000000u) >> 24) | \
-			   (((__u32)(x) & 0x00ff0000u) >>  8) | \
-			   (((__u32)(x) & 0x0000ff00u) <<  8) | \
-			   (((__u32)(x) & 0x000000ffu) << 24))
+#define bswap32(x) (uint32_t)	((((uint32_t)(x) & 0xff000000u) >> 24) | \
+				(((uint32_t)(x) & 0x00ff0000u) >>  8) | \
+				(((uint32_t)(x) & 0x0000ff00u) <<  8) | \
+				(((uint32_t)(x) & 0x000000ffu) << 24))
 
-#define bswap64(x) (__u64)((((__u64)(x) & 0xff00000000000000ull) >> 56) | \
-			   (((__u64)(x) & 0x00ff000000000000ull) >> 40) | \
-			   (((__u64)(x) & 0x0000ff0000000000ull) >> 24) | \
-			   (((__u64)(x) & 0x000000ff00000000ull) >>  8) | \
-			   (((__u64)(x) & 0x00000000ff000000ull) <<  8) | \
-			   (((__u64)(x) & 0x0000000000ff0000ull) << 24) | \
-			   (((__u64)(x) & 0x000000000000ff00ull) << 40) | \
-			   (((__u64)(x) & 0x00000000000000ffull) << 56))
+#define bswap64(x) (uint64_t)	((((uint64_t)(x) & 0xff00000000000000ull) >> 56) | \
+				(((uint64_t)(x) & 0x00ff000000000000ull) >> 40) | \
+				(((uint64_t)(x) & 0x0000ff0000000000ull) >> 24) | \
+				(((uint64_t)(x) & 0x000000ff00000000ull) >>  8) | \
+				(((uint64_t)(x) & 0x00000000ff000000ull) <<  8) | \
+				(((uint64_t)(x) & 0x0000000000ff0000ull) << 24) | \
+				(((uint64_t)(x) & 0x000000000000ff00ull) << 40) | \
+				(((uint64_t)(x) & 0x00000000000000ffull) << 56))
 
 #ifdef __BYTE_ORDER
 #if (__BYTE_ORDER == __LITTLE_ENDIAN)
@@ -87,14 +87,14 @@ enum endian {
 	BE = 1
 };
 
-extern void volume_id_set_unicode16(char *str, unsigned int len, const __u8 *buf, enum endian endianess, unsigned int count);
+extern void volume_id_set_unicode16(char *str, size_t len, const uint8_t *buf, enum endian endianess, size_t count);
 extern void volume_id_set_usage(struct volume_id *id, enum volume_id_usage usage_id);
 extern void volume_id_set_usage_part(struct volume_id_partition *part, enum volume_id_usage usage_id);
-extern void volume_id_set_label_raw(struct volume_id *id, const __u8 *buf, unsigned int count);
-extern void volume_id_set_label_string(struct volume_id *id, const __u8 *buf, unsigned int count);
-extern void volume_id_set_label_unicode16(struct volume_id *id, const __u8 *buf, enum endian endianess, unsigned int count);
-extern void volume_id_set_uuid(struct volume_id *id, const __u8 *buf, enum uuid_format format);
-extern __u8 *volume_id_get_buffer(struct volume_id *id, __u64 off, unsigned int len);
+extern void volume_id_set_label_raw(struct volume_id *id, const uint8_t *buf, size_t count);
+extern void volume_id_set_label_string(struct volume_id *id, const uint8_t *buf, size_t count);
+extern void volume_id_set_label_unicode16(struct volume_id *id, const uint8_t *buf, enum endian endianess, size_t count);
+extern void volume_id_set_uuid(struct volume_id *id, const uint8_t *buf, enum uuid_format format);
+extern uint8_t *volume_id_get_buffer(struct volume_id *id, uint64_t off, size_t len);
 extern void volume_id_free_buffer(struct volume_id *id);
 
 #endif /* _VOLUME_ID_UTIL_ */
